auth page cleanup
