--- web/app/(auth)/login/page.tsx
@@ -1,9 +1,0 @@
-import LoginComponent from "@/components/auth/Login";
-
-export default function LoginPage() {
-  return (
-    <section className="flex min-h-screen bg-zinc-50 px-4 py-16 md:py-32">
-      <LoginComponent />
-    </section>
-  );
-}
