--- web/app/(auth)/register/page.tsx
@@ -1,9 +1,0 @@
-import RegisterComponent from "@/components/auth/Register";
-
-export default function RegisterPage() {
-  return (
-    <section className="flex min-h-screen bg-zinc-50 px-4 py-16 md:py-32">
-      <RegisterComponent />
-    </section>
-  );
-}
